Switch site.css & blog_logo paths to relative root
I did this because the way they were currently implemented they were
explicitly templatesching the domain specified in the config. This isn't
always ideal. For instance lets say you want to test it out in a
different netlify branch based build or something. I believe this should
work correctly in all the normal cases as well.